Fix the check for controlled folder access
The controlled folder access feature can be set to at least three
states: disabled; enforcement mode; and audit mode. Only enforcement
mode actually presents a problem for SKIF. However, SKIF currently
issues a warning implying that controlled folder access is causing
a problem for SKIF even if the feature is only in audit mode.

This patch changes the check for controlled folder access so that
it is perceived as being enabled only if it is set to enforcement
mode (not audit mode).